backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1
gunicorn>=21.2.0

# OpenAI and Anthropic
openai>=1.12.0
anthropic>=0.8.1

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7

# File Handling
aiofiles>=23.2.1

# HTTP Client
requests>=2.31.0

# Web Scraping (if needed)
beautifulsoup4>=4.12.2
selenium>=4.16.0
webdriver-manager>=4.0.1

# PDF Generation
weasyprint>=60.1
jinja2>=3.1.2

# CORS
starlette>=0.36.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0

# Development
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0 